+++ src/fhirpath/dialects/elasticsearch.py
@@ -97,17 +97,36 @@ class ElasticSearchDialect(DialectBase):
             if len(value) == 1:
                 value = value[0]
             else:
-                q = {
-                    "bool": {"should": [], "minimum_should_match": 1}
-                }
+                q = {"bool": {"should": [], "minimum_should_match": 1}}
                 for val in value:
-                    q["bool"]["should"].append(
-                        {"prefix": {path: {"value": val}}}
-                    )
+                    q["bool"]["should"].append({"prefix": {path: {"value": val}}})
                 return q
 
         q = {"prefix": {path: {"value": value}}}
-        # xxx: what about multiple
+
+        return q
+
+    def _create_eb_term(self, path, value):
+        """Create ES Prefix Query"""
+
+        def _check(v):
+            if "*" in v:
+                raise NotImplementedError
+
+        if isinstance(value, (list, tuple)):
+            if len(value) == 1:
+                value = value[0]
+            else:
+                q = {"bool": {"should": [], "minimum_should_match": 1}}
+                for val in value:
+                    _check(val)
+                    q["bool"]["should"].append(
+                        {"wildcard": {path: {"value": "*" + val}}}
+                    )
+                return q
+        _check(value)
+        q = {"wildcard": {path: {"value": "*" + value}}}
+
         return q
 
     def _create_dotted_path(self, term, root_replacer=None):
@@ -303,9 +322,10 @@ class ElasticSearchDialect(DialectBase):
 
                     else:
                         if term.comparison_operator == OPERATOR.sa:
-                            q = self._create_sa_term(
-                                dotted_path, value
-                            )
+                            q = self._create_sa_term(dotted_path, value)
+                        elif term.comparison_operator == OPERATOR.eb:
+                            q = self._create_eb_term(dotted_path, value)
+
                         else:
                             q = self._create_term(dotted_path, value, multiple=multiple)
                         resolved = q, term.unary_operator
@@ -459,6 +479,8 @@ class ElasticSearchDialect(DialectBase):
                 qr = {"match_phrase": {path_: value}}
             elif term.comparison_operator == OPERATOR.sa:
                 qr = {"match_phrase_prefix": {path_: value}}
+            elif term.comparison_operator == OPERATOR.eb:
+                qr = {"match": {path_: {"query": value, "operator": "AND"}}}
             else:
                 qr = {"match": {path_: value}}
         elif ("/" in value or URI_SCHEME.match(value)) and ".reference" in path_: