+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
 Improvements
 
+- Add support for ``:above``FHIR search modifier and `èb``prefix. See https://github.com/nazrulworld/fhirpath/issues/2
+
 - Add support for ``:bellow`` FHIR search modifier and ``sa`` prefix. See https://github.com/nazrulworld/fhirpath/issues/2
 
 Bugfixes